move to self test controller
