+++ VideoWeb/VideoWeb.UnitTests/VideoWeb.UnitTests.csproj
@@ -32,4 +32,7 @@
     <ProjectReference Include="..\VideoWeb.Common\VideoWeb.Common.csproj" />
     <ProjectReference Include="..\VideoWeb\VideoWeb.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Controllers\TestCallsController\" />
+  </ItemGroup>
 </Project>
