+++ VideoWeb/VideoWeb/Controllers/SelfTestController.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Net;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
@@ -7,6 +8,8 @@ using VideoWeb.Mappings;
 using VideoApi.Client;
 using VideoApi.Contract.Responses;
 using System.Threading.Tasks;
+using VideoWeb.Common.Caching;
+using VideoWeb.Middleware;
 
 namespace VideoWeb.Controllers
 {
@@ -16,14 +19,16 @@ namespace VideoWeb.Controllers
     public class SelfTestController : Controller
     {
         private readonly IVideoApiClient _videoApiClient;
+        private readonly ITestCallCache _testCallCache;
         private readonly ILogger<SelfTestController> _logger;
         private readonly IMapperFactory _mapperFactory;
 
-        public SelfTestController(IVideoApiClient videoApiClient, ILogger<SelfTestController> logger, IMapperFactory mapperFactory)
+        public SelfTestController(IVideoApiClient videoApiClient, ILogger<SelfTestController> logger, IMapperFactory mapperFactory, ITestCallCache testCallCache)
         {
             _videoApiClient = videoApiClient;
             _logger = logger;
             _mapperFactory = mapperFactory;
+            _testCallCache = testCallCache;
         }
 
         /// <summary>
@@ -47,6 +52,62 @@ namespace VideoWeb.Controllers
             catch (VideoApiException e)
             {
                 _logger.LogError(e, $"Unable to get Pexip configuration");
+                return StatusCode(e.StatusCode, e.Response);
+            }
+        }
+        
+        /// <summary>
+        /// Check if a user has completed a test call at least once today
+        /// </summary>
+        /// <returns>OK if a test has been completed at least once, else not found</returns>
+        [HttpGet("today")]
+        [SwaggerOperation(OperationId = "CheckUserCompletedATestToday")]
+        [ProducesResponseType((int)HttpStatusCode.OK)]
+        [ProducesResponseType((int)HttpStatusCode.NotFound)]
+        public async Task<IActionResult> CheckUserCompletedATestTodayAsync()
+        {
+            var hasUserCompletedATestToday = await _testCallCache.HasUserCompletedATestToday(User.Identity.Name);
+            return hasUserCompletedATestToday ? Ok() : NotFound();
+        }
+        
+        [ServiceFilter(typeof(CheckParticipantCanAccessConferenceAttribute))]
+        [HttpGet]
+        [Route("conferences/{conferenceId}/participants/{participantId}/selftestresult")]
+        [SwaggerOperation(OperationId = "GetTestCallResult")]
+        [ProducesResponseType(typeof(TestCallScoreResponse), (int)HttpStatusCode.OK)]
+        [ProducesResponseType((int)HttpStatusCode.NotFound)]
+        public async Task<IActionResult> GetTestCallResultForParticipantAsync(Guid conferenceId, Guid participantId)
+        {
+            try
+            {
+                var score = await _videoApiClient.GetTestCallResultForParticipantAsync(conferenceId, participantId);
+                await _testCallCache.AddTestCompletedForTodayAsync(User.Identity.Name);
+                return Ok(score);
+            }
+            catch (VideoApiException e)
+            {
+                _logger.LogError(e,
+                    "Unable to get test call result for participant: {ParticipantId} in conference: {ConferenceId}",
+                    participantId, conferenceId);
+                return StatusCode(e.StatusCode, e.Response);
+            }
+        }
+        
+        [HttpGet("independentselftestresult/{participantId}")]
+        [SwaggerOperation(OperationId = "GetIndependentTestCallResult")]
+        [ProducesResponseType(typeof(TestCallScoreResponse), (int)HttpStatusCode.OK)]
+        [ProducesResponseType((int)HttpStatusCode.NotFound)]
+        public async Task<IActionResult> GetIndependentTestCallResultAsync(Guid participantId)
+        {
+            try
+            {
+                var score = await _videoApiClient.GetIndependentTestCallResultAsync(participantId);
+                await _testCallCache.AddTestCompletedForTodayAsync(User.Identity.Name);
+                return Ok(score);
+            }
+            catch (VideoApiException e)
+            {
+                _logger.LogError(e, $"Unable to get independent test call result for participant: {participantId}");
                 return StatusCode(e.StatusCode, e.Response);
             }
         }
